Update Culling and Draw Ranges Jobs, add new reference, bump version
Added a new GUID reference in BrgContainer.Runtime.asmdef. In the CullingBatchInstancesJob, the way object position and culling are being processed is updated to use AABB rather than individual points, optimizing the culling process. The DrawRangesJob in BatchRendererGroupContainer has been slightly updated to remove a trailing comma. Finally, the version of the package has been bumped to 1.2.2.