Refactor LOD selection and copying visibility indices
Refactored the SelectLodPerInstanceJob to adhere to IJobFilter and also modified its execution control flow. The change includes an update in the execution control flow where we check for level of detail (LOD) and early return if the condition is met. In addition, renamed variables in BatchLodDescription.cs for more clarity. Also, improved the copying of visibility indices in CopyVisibilityIndicesToArrayJob.cs by adjusting the pointers and freeing memory appropriately. Lastly, updated the BatchRendererGroupContainer.cs to reflect these changes.